% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{runPathwayVolcano}
\alias{runPathwayVolcano}
\title{Launch the PathwayVolcano Shiny App}
\usage{
runPathwayVolcano()
}
\description{
This function launches the interactive Shiny app for exploring pathway-specific volcano plots.
}
